Add track moving to the list level
This way we can start to respond to the difference between being in the list and outside.
