+++ js/components/PlaylistWindow/TrackList.js
@@ -3,27 +3,65 @@ import { connect } from "react-redux";
 
 import { getTimeStr } from "../../utils";
 import { getVisibleTrackIds, getScrollOffset } from "../../selectors";
+import { TRACK_HEIGHT } from "../../constants";
 import { SELECT_ZERO } from "../../actionTypes";
+import { dragSelected } from "../../actionCreators";
 import TrackCell from "./TrackCell";
 import TrackTitle from "./TrackTitle";
 
 class TrackList extends React.Component {
   constructor(props) {
     super(props);
+    this._handleMoveClick = this._handleMoveClick.bind(this);
   }
 
   _renderTracks(format) {
     return this.props.trackIds.map((id, i) => (
-      <TrackCell key={id} id={id} index={i}>
+      <TrackCell
+        key={id}
+        id={id}
+        index={i}
+        handleMoveClick={this._handleMoveClick}
+      >
         {format(id, i)}
       </TrackCell>
     ));
+  }
+
+  _handleMoveClick(e) {
+    if (!this._node) {
+      return;
+    }
+    const { top, bottom, left, right } = this._node.getBoundingClientRect();
+    const mouseStart = e.clientY;
+    let lastDiff = 0;
+    const handleMouseMove = ee => {
+      const { clientY: y, clientX: x } = ee;
+      if (y < top || y > bottom || x < left || x > right) {
+        // Mouse is outside the track list
+        return;
+      }
+      const proposedDiff = Math.floor((y - mouseStart) / TRACK_HEIGHT);
+      if (proposedDiff !== lastDiff) {
+        const diffDiff = proposedDiff - lastDiff;
+        this.props.dragSelected(diffDiff);
+        lastDiff = proposedDiff;
+      }
+    };
+
+    window.addEventListener("mouseup", () => {
+      window.removeEventListener("mousemove", handleMouseMove);
+    });
+    window.addEventListener("mousemove", handleMouseMove);
   }
 
   render() {
     const { tracks, offset } = this.props;
     return (
       <div
+        ref={node => {
+          this._node = node;
+        }}
         className="playlist-tracks"
         style={{ height: "100%" }}
         onClick={this.props.selectZero}
@@ -43,7 +81,8 @@ class TrackList extends React.Component {
 }
 
 const mapDispatchToProps = {
-  selectZero: () => ({ type: SELECT_ZERO })
+  selectZero: () => ({ type: SELECT_ZERO }),
+  dragSelected
 };
 
 const mapStateToProps = state => ({
